pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "retail-sql-agent"
version = "0.1.0"
description = "A self-correcting text-to-SQL agent for a retail analytics SQLite database using modern AutoGen."
readme = "README.md"
requires-python = ">=3.11"
authors = [{ name = "Sai" }]
dependencies = [
    "autogen-agentchat>=0.7.2",
    "autogen-ext[openai]>=0.7.2",
    "python-dotenv>=1.0.1",
    "pydantic>=2.7.0",
    "streamlit>=1.44.0",
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
